Move media service to SavedMusic package
