+++ app/src/main/java/com/yurii/youtubemusic/screens/saved/service/MusicServiceConnection.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.services.mediaservice
+package com.yurii.youtubemusic.screens.saved.service
 
 import android.content.ComponentName
 import android.content.Context
